commerce-admin/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.3
certifi==2020.6.20
chardet==3.0.4
Django==3.0.6
django-cors-headers==3.4.0
django-environ==0.4.5
django-extensions==2.2.9
django-polymorphic==2.1.2
django-redis==4.12.1
django-rest-polymorphic==0.1.9
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.9
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-file==1.5.1
six==1.15.0
sqlparse==0.3.1
tldextract==2.2.2
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
